+++ pytabulator/__init__.py
@@ -1,6 +1,7 @@
 # from ._table_options_dc import TableOptions
+from ._table_options_pydantic import TableOptions
 from .shiny_bindings import output_tabulator, render_data_frame, render_tabulator
-from .tabulator import TableOptions, Tabulator
+from .tabulator import Tabulator
 from .tabulator_context import TabulatorContext
 
 # __all__ = []
